Thread safety improvements in DT EventSetup formats
Two singletons are removed. The content that was in
DTCompactMapPluginHandler was moved into the fullMap
function of DTReadoutMapping. The content that was in
DTConfigHandler was moved into the new class
DTKeyedConfigCache. The two modules that used it now
have that as a data member instead of accessing the
functionality through a singleton. In both cases,
the algorithmic code was simply copied to the new
location with minimal changes.

Note the way DTConfigHandler used KeyList with a const_cast
was a problem and remains a problem. I did not fix that
problem. I just noticed it and copied that code from one
place to another. This will probably need to be fixed
at some point in the future.

Remove an optimization that was not needed from
DTReadOutMapping that was added in a very recent
commit by me. Recently I added the atomicCache_ and
cache_ data members and only the atomicCache_ is
really needed so I removed cache_ to simplify the code.

Removed two lines of code that did nothing and some
functions that were never called by anything.
And I renamed a local variable from theCache to
localCache for clarity.
