+++ CondCore/DTPlugins/interface/DTConfigPluginHandler.h
@@ -1,81 +1,1 @@
-#ifndef DTConfigPluginHandler_H
-#define DTConfigPluginHandler_H
-/** \class DTConfigPluginHandler
- *
- *  Description:
- *       Class to hold configuration identifier for chambers
- *
- *  $Date: 2010/05/14 11:43:02 $
- *  $Revision: 1.2 $
- *  \author Paolo Ronchese INFN Padova
- *
- */
-
-//----------------------
-// Base Class Headers --
-//----------------------
-#include "CondFormats/DTObjects/interface/DTConfigAbstractHandler.h"
-
-//------------------------------------
-// Collaborating Class Declarations --
-//------------------------------------
-
-
-//---------------
-// C++ Headers --
-//---------------
-#include <string>
-#include <map>
-
-//              ---------------------
-//              -- Class Interface --
-//              ---------------------
-
-class DTConfigPluginHandler: public DTConfigAbstractHandler {
-
- public:
-
-  /** Destructor
-   */
-  virtual ~DTConfigPluginHandler();
-
-  /** Operations
-   */
-  /// build static object
-  static void build();
-
-  /// get content
-  virtual int get( const edm::EventSetup& context,
-                   int cfgId, const DTKeyedConfig*& obj );
-  virtual int get( const DTKeyedConfigListRcd& keyRecord,
-                   int cfgId, const DTKeyedConfig*& obj );
-  virtual void getData( const edm::EventSetup& context,
-                        int cfgId, std::vector<std::string>& list );
-  virtual void getData( const DTKeyedConfigListRcd& keyRecord,
-                        int cfgId, std::vector<std::string>& list );
-
-  void purge();
-
-  static const int maxBrickNumber;
-  static const int maxStringNumber;
-  static const int maxByteNumber;
-
- private:
-
-  /** Constructor
-   */
-  DTConfigPluginHandler();
-  DTConfigPluginHandler( const DTConfigPluginHandler& x );
-  const DTConfigPluginHandler& operator=( const DTConfigPluginHandler& x );
-
-  typedef std::pair<int,const DTKeyedConfig*> counted_brick;
-  std::map<int,counted_brick> brickMap;
-  int cachedBrickNumber;
-  int cachedStringNumber;
-  int cachedByteNumber;
-
-};
-
-#endif // DTConfigPluginHandler_H
-
-
+#error obsolete header
