--- CondCore/DTPlugins/src/DTCompactMapHandlerFactory.cc
@@ -1,15 +1,0 @@
-
-#include "CondCore/DTPlugins/interface/DTCompactMapPluginHandler.h"
-
-class DTCompactMapHandlerFactory {
-
- public:
-
-  DTCompactMapHandlerFactory() {
-    DTCompactMapPluginHandler::build();
-  }
-
-};
-
-static DTCompactMapHandlerFactory dtcmhf;
-
