--- CondCore/DTPlugins/src/DTConfigHandlerFactory.cc
@@ -1,15 +1,0 @@
-
-#include "CondCore/DTPlugins/interface/DTConfigPluginHandler.h"
-
-class DTConfigHandlerFactory {
-
- public:
-
-  DTConfigHandlerFactory() {
-    DTConfigPluginHandler::build();
-  }
-
-};
-
-static const DTConfigHandlerFactory dtchf;
-
