+++ CondFormats/DTObjects/interface/DTReadOutMapping.h
@@ -149,8 +149,6 @@ class DTReadOutMapping {
   /// Expand to full map
   const DTReadOutMapping* fullMap() const;
 
-  void initialize();
-
  private:
 
   DTReadOutMapping(DTReadOutMapping const&);
@@ -159,40 +157,13 @@ class DTReadOutMapping {
   edm::AtomicPtrCache<DTReadOutMappingCache> const& atomicCache() const { return atomicCache_; }
   edm::AtomicPtrCache<DTReadOutMappingCache> & atomicCache() { return atomicCache_; }
 
-  edm::ConstRespectingPtr<DTReadOutMappingCache> const& cache() const { return cache_; }
-  edm::ConstRespectingPtr<DTReadOutMappingCache> & cache() { return cache_; }
+  static DTReadOutMapping* expandMap( const DTReadOutMapping& compMap );
 
   std::string cellMapVersion;
   std::string  robMapVersion;
 
   std::vector<DTReadOutGeometryLink> readOutChannelDriftTubeMap;
 
-  // There are some caches to help look up the data in the
-  // preceding vector. cache_ holds a pointer to several
-  // maps. Normally it is automatically filled immediately
-  // after the object is read in from the database by the
-  // initialize function.  The initialize function is a
-  // non const function and it is not safe to call it
-  // concurrently.  That is why atomicCache_ exists.
-  // It holds exactly the same information, but the function
-  // that fills it is declared const and can be called concurrently.
-  // When the functions that use the caches are called
-  // the first time, atomicCache_ is filled if cache_
-  // has not already been filled. The initialize function
-  // is implemented to fill atomicCache_
-  // if it is not already filled, then move the pointer
-  // it holds into cache_. One would use atomicCache_
-  // in cases where the object is not read in from the
-  // database and the insert function was used to fill
-  // it. After all the inserts are done one could call
-  // readoutToGeometry or its inverse and use the maps.
-  // With the plain (noncompact) format, one can also do
-  // that even before all the inserts are done.
-  // rgBuf and grBuf are filled as new entries are inserted
-  // in the vector before either cache was filled.
-  // The caches contain their own rgBuf and grBuf after they
-  // are filled.
-  edm::ConstRespectingPtr<DTReadOutMappingCache> cache_;
   edm::AtomicPtrCache<DTReadOutMappingCache> atomicCache_;
   edm::ConstRespectingPtr<DTBufferTree<int,int> > rgBuf;
   edm::ConstRespectingPtr<DTBufferTree<int,int> > grBuf;
