+++ CondFormats/DTObjects/src/DTReadOutMapping.cc
@@ -14,7 +14,6 @@
 // Collaborating Class Headers --
 //-------------------------------
 #include "CondFormats/DTObjects/interface/DTBufferTree.h"
-#include "CondFormats/DTObjects/interface/DTCompactMapAbstractHandler.h"
 #include "CondFormats/DTObjects/interface/DTReadOutMappingCache.h"
 
 //---------------
@@ -130,31 +129,12 @@ int DTReadOutMapping::readOutToGeometry( int      dduId,
   layerId   =
   cellId    = 0;
 
-  DTBufferTree<int,int> const* pmType;
-  DTBufferTree<int,int> const* prgBuf;
-  DTBufferTree<int,int> const* prgDDU;
-  DTBufferTree<int,int> const* prgROS;
-  DTBufferTree<int,int> const* prgROB;
-
-  if(cache().isSet()) {
-    pmType = cache()->mType.get();
-    prgBuf = cache()->rgBuf.get();
-    prgDDU = cache()->rgDDU.get();
-    prgROS = cache()->rgROS.get();
-    prgROB = cache()->rgROB.get();
-  } else {
-    if(!atomicCache().isSet()) {
-      cacheMap();
-    }
-    pmType = atomicCache()->mType.get();
-    prgBuf = atomicCache()->rgBuf.get();
-    prgDDU = atomicCache()->rgDDU.get();
-    prgROS = atomicCache()->rgROS.get();
-    prgROB = atomicCache()->rgROB.get();
+  if(!atomicCache().isSet()) {
+    cacheMap();
   }
 
   int defaultValue;
-  pmType->find( 0, defaultValue );
+  atomicCache()->mType.find( 0, defaultValue );
   if ( defaultValue ) {
 
     int searchStatus;
@@ -164,7 +144,7 @@ int DTReadOutMapping::readOutToGeometry( int      dduId,
     dduKey.reserve( 5 );
     dduKey.push_back( dduId );
     dduKey.push_back( rosId );
-    searchStatus = prgDDU->find( dduKey.begin(), dduKey.end(), ientry );
+    searchStatus = atomicCache()->rgDDU.find( dduKey.begin(), dduKey.end(), ientry );
     if ( searchStatus ) return searchStatus;
     const DTReadOutGeometryLink& lros( readOutChannelDriftTubeMap[ientry] );
     wheelId = lros.  wheelId;
@@ -174,7 +154,7 @@ int DTReadOutMapping::readOutToGeometry( int      dduId,
     rosKey.reserve( 5 );
     rosKey.push_back( lros. cellId );
     rosKey.push_back( robId );
-    searchStatus = prgROS->find( rosKey.begin(), rosKey.end(), ientry );
+    searchStatus = atomicCache()->rgROS.find( rosKey.begin(), rosKey.end(), ientry );
     if ( searchStatus ) return searchStatus;
     const DTReadOutGeometryLink& lrob( readOutChannelDriftTubeMap[ientry] );
     if ( lrob.  wheelId != defaultValue )  wheelId = lrob.  wheelId;
@@ -186,7 +166,7 @@ int DTReadOutMapping::readOutToGeometry( int      dduId,
     robKey.push_back( lrob. cellId );
     robKey.push_back(     tdcId );
     robKey.push_back( channelId );
-    searchStatus = prgROB->find( robKey.begin(), robKey.end(), ientry );
+    searchStatus = atomicCache()->rgROB.find( robKey.begin(), robKey.end(), ientry );
     if ( searchStatus ) return searchStatus;
     const DTReadOutGeometryLink& ltdc( readOutChannelDriftTubeMap[ientry] );
          slId = ltdc.     slId;
@@ -204,7 +184,7 @@ int DTReadOutMapping::readOutToGeometry( int      dduId,
   chanKey.push_back(     tdcId );
   chanKey.push_back( channelId );
   int ientry;
-  int searchStatus = prgBuf->find( chanKey.begin(), chanKey.end(), ientry );
+  int searchStatus = atomicCache()->rgBuf.find( chanKey.begin(), chanKey.end(), ientry );
   if ( !searchStatus ) {
     const DTReadOutGeometryLink& link( readOutChannelDriftTubeMap[ientry] );
       wheelId = link.  wheelId;
@@ -258,31 +238,12 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
   tdcId =
   channelId = 0;
 
-  DTBufferTree<int,int> const* pmType;
-  DTBufferTree<int,int> const* pgrBuf;
-  DTBufferTreeUniquePtr const* pgrDDU;
-  DTBufferTreeUniquePtr const* pgrROS;
-  DTBufferTreeUniquePtr const* pgrROB;
-
-  if(cache().isSet()) {
-    pmType = cache()->mType.get();
-    pgrBuf = cache()->grBuf.get();
-    pgrDDU = cache()->grDDU.get();
-    pgrROS = cache()->grROS.get();
-    pgrROB = cache()->grROB.get();
-  } else {
-    if(!atomicCache().isSet()) {
-      cacheMap();
-    }
-    pmType = atomicCache()->mType.get();
-    pgrBuf = atomicCache()->grBuf.get();
-    pgrDDU = atomicCache()->grDDU.get();
-    pgrROS = atomicCache()->grROS.get();
-    pgrROB = atomicCache()->grROB.get();
+  if(!atomicCache().isSet()) {
+    cacheMap();
   }
 
   int defaultValue;
-  pmType->find( 0, defaultValue );
+  atomicCache()->mType.find( 0, defaultValue );
   if ( defaultValue ) {
 
     int loop1 = 0;
@@ -302,7 +263,7 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
     cellKey.push_back( layerId );
     cellKey.push_back(    slId );
     std::vector<int> stdcKey = cellKey;
-    searchStatus = pgrROB->find( cellKey.begin(), cellKey.end(), robMLgr );
+    searchStatus = atomicCache()->grROB.find( cellKey.begin(), cellKey.end(), robMLgr );
     if ( searchStatus ) return searchStatus;
     if ( !( robMLgr->size() ) ) return 1;
     std::vector<int>::const_iterator tdc_iter = robMLgr->begin();
@@ -318,7 +279,7 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
       cellKey.push_back( mapId );
       cellKey.push_back( stationId );
       std::vector<int> srosKey = cellKey;
-      searchStatus = pgrROS->find( cellKey.begin(), cellKey.end(), rosMLgr );
+      searchStatus = atomicCache()->grROS.find( cellKey.begin(), cellKey.end(), rosMLgr );
       if ( searchStatus ) continue;
       if ( !( rosMLgr->size() ) ) continue;
       std::vector<int>::const_iterator ros_iter = rosMLgr->begin();
@@ -340,7 +301,7 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
         cellKey.push_back(  wheelId );
         cellKey.push_back( sectorId );
         std::vector<int> sdduKey = cellKey;
-        searchStatus = pgrDDU->find( cellKey.begin(), cellKey.end(), dduMLgr );
+        searchStatus = atomicCache()->grDDU.find( cellKey.begin(), cellKey.end(), dduMLgr );
         if ( searchStatus ) continue;
         if ( !( dduMLgr->size() ) ) continue;
         if ( searchStatus ) return searchStatus;
@@ -352,7 +313,6 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
           loop3++;
           const DTReadOutGeometryLink& lddu(
                 readOutChannelDriftTubeMap[*ddu_iter++] );
-          mapId = lros.rosId;
           if ( ( ( sectorId == secCk ) || 
                  ( sectorId == lddu.sectorId ) ) &&
                ( (  wheelId == wheCk ) || 
@@ -378,7 +338,7 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
   cellKey.push_back(   layerId );
   cellKey.push_back(    cellId );
   int ientry;
-  int searchStatus = pgrBuf->find( cellKey.begin(), cellKey.end(), ientry );
+  int searchStatus = atomicCache()->grBuf.find( cellKey.begin(), cellKey.end(), ientry );
   if ( !searchStatus ) {
     const DTReadOutGeometryLink& link( readOutChannelDriftTubeMap[ientry] );
         dduId = link.    dduId;
@@ -395,18 +355,12 @@ int DTReadOutMapping::geometryToReadOut( int    wheelId,
 
 DTReadOutMapping::type DTReadOutMapping::mapType() const {
 
-  DTBufferTree<int,int> const* pmType;
-  if(cache().isSet()) {
-    pmType = cache()->mType.get();
-  } else {
-    if(!atomicCache().isSet()) {
-      cacheMap();
-    }
-    pmType = atomicCache()->mType.get();
+  if(!atomicCache().isSet()) {
+    cacheMap();
   }
 
   int defaultValue;
-  pmType->find( 0, defaultValue );
+  atomicCache()->mType.find( 0, defaultValue );
   if ( defaultValue ) return compact;
   else                return plain;
 }
@@ -435,7 +389,6 @@ std::string& DTReadOutMapping::mapRobRos() {
 
 
 void DTReadOutMapping::clear() {
-  cache().reset();
   atomicCache().reset();
   rgBuf->clear();
   grBuf->clear();
@@ -474,12 +427,9 @@ int DTReadOutMapping::insertReadOutGeometryLink( int     dduId,
   DTBufferTree<int,int>* pgrBuf;
   DTBufferTree<int,int>* prgBuf;
 
-  if(cache().isSet()) {
-    pgrBuf = cache()->grBuf.get();
-    prgBuf = cache()->rgBuf.get();
-  } else if(atomicCache().isSet()) {
-    pgrBuf = atomicCache()->grBuf.get();
-    prgBuf = atomicCache()->rgBuf.get();
+  if(atomicCache().isSet()) {
+    pgrBuf = &atomicCache()->grBuf;
+    prgBuf = &atomicCache()->rgBuf;
   } else {
     pgrBuf = grBuf.get();
     prgBuf = rgBuf.get();
@@ -522,28 +472,96 @@ DTReadOutMapping::const_iterator DTReadOutMapping::end() const {
 
 const DTReadOutMapping* DTReadOutMapping::fullMap() const {
   if ( mapType() == plain ) return this;
-  DTCompactMapAbstractHandler* cmHandler =
-                               DTCompactMapAbstractHandler::getInstance();
-  if ( cmHandler == 0 ) {
-      std::cout << "CondCoreDTPlugins library not loaded, "
-                << "compactMapHandler plugin missing" << std::endl;
-    return 0;
-  }
-  return cmHandler->expandMap( *this );
+  return expandMap( *this );
 }
 
 
-void DTReadOutMapping::initialize() {
-  if(cache().isSet()) {
-    return;
-  } else if(atomicCache().isSet()) {
-    cache().set(std::unique_ptr<DTReadOutMappingCache>(atomicCache().release()));
-  } else {
-    cacheMap();
-    cache().set(std::unique_ptr<DTReadOutMappingCache>(atomicCache().release()));
-  }
-}
+// The code for this function was copied verbatim from
+// CondCore/DTPlugins/src/DTCompactMapPluginHandler.c
+DTReadOutMapping* DTReadOutMapping::expandMap( const DTReadOutMapping& compMap ) {
+  std::vector<DTReadOutGeometryLink> entryList;
+  DTReadOutMapping::const_iterator compIter = compMap.begin();
+  DTReadOutMapping::const_iterator compIend = compMap.end();
+  while ( compIter != compIend ) entryList.push_back( *compIter++ );
 
+  std::string
+  rosMap = "expand_";
+  rosMap += compMap.mapRobRos();
+  std::string
+  tdcMap = "expand_";
+  tdcMap += compMap.mapCellTdc();
+  DTReadOutMapping* fullMap = new DTReadOutMapping( tdcMap, rosMap );
+  int ddu;
+  int ros;
+  int rch;
+  int tdc;
+  int tch;
+  int whe;
+  int sta;
+  int sec;
+  int rob;
+  int qua;
+  int lay;
+  int cel;
+  int mt1;
+  int mi1;
+  int mt2;
+  int mi2;
+  int def;
+  int wha;
+  int sea;
+  std::vector<DTReadOutGeometryLink>::const_iterator iter = entryList.begin();
+  std::vector<DTReadOutGeometryLink>::const_iterator iend = entryList.end();
+  std::vector<DTReadOutGeometryLink>::const_iterator iros = entryList.end();
+  std::vector<DTReadOutGeometryLink>::const_iterator irob = entryList.end();
+  while ( iter != iend ) {
+    const DTReadOutGeometryLink& rosEntry( *iter++ );
+    if ( rosEntry.dduId > 0x3fffffff ) continue;
+    ddu = rosEntry.dduId;
+    ros = rosEntry.rosId;
+    whe = rosEntry.wheelId;
+    def = rosEntry.stationId;
+    sec = rosEntry.sectorId;
+    rob = rosEntry.slId;
+    mt1 = rosEntry.layerId;
+    mi1 = rosEntry.cellId;
+    iros = entryList.begin();
+    while ( iros != iend ) {
+      wha = whe;
+      sea = sec;
+      const DTReadOutGeometryLink& rchEntry( *iros++ );
+      if ( ( rchEntry.dduId != mt1 ) ||
+           ( rchEntry.rosId != mi1 ) ) continue;
+      rch =  rchEntry.robId;
+      if (   rchEntry.wheelId != def   ) wha = rchEntry.wheelId;
+      sta =  rchEntry.stationId;
+      if (   rchEntry.sectorId != def   ) sea = rchEntry.sectorId;
+      rob =  rchEntry.slId;
+      mt2 =  rchEntry.layerId;
+      mi2 =  rchEntry.cellId;
+      irob = entryList.begin();
+      while ( irob != iend ) {
+        const DTReadOutGeometryLink& robEntry( *irob++ );
+        if ( ( robEntry.dduId != mt2 ) ||
+             ( robEntry.rosId != mi2 ) ) continue;
+        if (   robEntry.robId != rob   ) {
+          std::cout << "ROB mismatch " << rob << " "
+                                       << robEntry.robId << std::endl;
+        }
+        tdc =  robEntry.tdcId;
+        tch =  robEntry.channelId;
+        qua =  robEntry.slId;
+        lay =  robEntry.layerId;
+        cel =  robEntry.cellId;
+        fullMap->insertReadOutGeometryLink( ddu, ros, rch, tdc, tch,
+                                                 wha, sta, sea,
+                                                 qua, lay, cel );
+
+      }
+    }
+  }
+  return fullMap;
+}
 
 std::string DTReadOutMapping::mapNameGR() const {
   std::stringstream name;
@@ -561,21 +579,9 @@ std::string DTReadOutMapping::mapNameRG() const {
 
 void DTReadOutMapping::cacheMap() const {
 
-  std::unique_ptr<DTReadOutMappingCache> theCache(new DTReadOutMappingCache);
+  std::unique_ptr<DTReadOutMappingCache> localCache(new DTReadOutMappingCache);
 
-  theCache->mType.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-
-  theCache->mType->insert( 0, 0 );
-
-  theCache->rgBuf.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-  theCache->rgROB.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-  theCache->rgROS.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-  theCache->rgDDU.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-  theCache->grBuf.set(std::unique_ptr<DTBufferTree<int,int> >(new DTBufferTree<int,int>));
-
-  theCache->grROB.set(std::unique_ptr<DTBufferTreeUniquePtr>(new DTBufferTreeUniquePtr));
-  theCache->grROS.set(std::unique_ptr<DTBufferTreeUniquePtr>(new DTBufferTreeUniquePtr));
-  theCache->grDDU.set(std::unique_ptr<DTBufferTreeUniquePtr>(new DTBufferTreeUniquePtr));
+  localCache->mType.insert( 0, 0 );
 
   int entryNum = 0;
   int entryMax = readOutChannelDriftTubeMap.size();
@@ -597,11 +603,11 @@ void DTReadOutMapping::cacheMap() const {
     val = link.stationId;
     if ( key > 0x3fffffff ) {
       if ( link.  tdcId > 0x3fffffff ) {
-        theCache->mType->insert( 0, defaultValue = link.  tdcId );
+        localCache->mType.insert( 0, defaultValue = link.  tdcId );
         rosMapKey = key;
       }
       else {
-        theCache->mType->insert( 0, defaultValue = link.wheelId );
+        localCache->mType.insert( 0, defaultValue = link.wheelId );
         robMapKey = key;
       }
     }
@@ -615,7 +621,7 @@ void DTReadOutMapping::cacheMap() const {
       chanKey.push_back( link.    tdcId );
       chanKey.push_back( link.channelId );
 
-      theCache->rgBuf->insert( chanKey.begin(), chanKey.end(), entryNum );
+      localCache->rgBuf.insert( chanKey.begin(), chanKey.end(), entryNum );
 
       cellKey.clear();
       cellKey.push_back( link.  wheelId );
@@ -625,7 +631,7 @@ void DTReadOutMapping::cacheMap() const {
       cellKey.push_back( link.  layerId );
       cellKey.push_back( link.   cellId );
 
-      theCache->grBuf->insert( cellKey.begin(), cellKey.end(), entryNum );
+      localCache->grBuf.insert( cellKey.begin(), cellKey.end(), entryNum );
 
     }
 
@@ -635,19 +641,19 @@ void DTReadOutMapping::cacheMap() const {
       chanKey.push_back( link.    rosId );
       chanKey.push_back( link.    tdcId );
       chanKey.push_back( link.channelId );
-      theCache->rgROB->insert( chanKey.begin(), chanKey.end(), entryNum );
+      localCache->rgROB.insert( chanKey.begin(), chanKey.end(), entryNum );
 
       cellKey.clear();
       cellKey.push_back( link. cellId );
       cellKey.push_back( link.layerId );
       cellKey.push_back( link.   slId );
       std::vector<int>* robMLgr;
-      theCache->grROB->find( cellKey.begin(), cellKey.end(), robMLgr );
+      localCache->grROB.find( cellKey.begin(), cellKey.end(), robMLgr );
       if ( robMLgr == 0 ) {
         std::unique_ptr<std::vector<int> > newVector(new std::vector<int>);
         robMLgr = newVector.get();
-        theCache->grROB->insert( cellKey.begin(), cellKey.end(),
-                                 std::move(newVector));
+        localCache->grROB.insert( cellKey.begin(), cellKey.end(),
+                                  std::move(newVector));
       }
       robMLgr->push_back( entryNum );
     }
@@ -657,18 +663,18 @@ void DTReadOutMapping::cacheMap() const {
       chanKey.clear();
       chanKey.push_back( link.rosId );
       chanKey.push_back( link.robId );
-      theCache->rgROS->insert( chanKey.begin(), chanKey.end(), entryNum );
+      localCache->rgROS.insert( chanKey.begin(), chanKey.end(), entryNum );
 
       cellKey.clear();
       cellKey.push_back( link.   cellId );
       cellKey.push_back( link.stationId );
       std::vector<int>* rosMLgr;
-      theCache->grROS->find( cellKey.begin(), cellKey.end(), rosMLgr );
+      localCache->grROS.find( cellKey.begin(), cellKey.end(), rosMLgr );
       if ( rosMLgr == 0 ) {
         std::unique_ptr<std::vector<int> > newVector(new std::vector<int>);
         rosMLgr = newVector.get();
-        theCache->grROS->insert( cellKey.begin(), cellKey.end(),
-                                 std::move(newVector));
+        localCache->grROS.insert( cellKey.begin(), cellKey.end(),
+                                  std::move(newVector));
       }
       rosMLgr->push_back( entryNum );
     }
@@ -679,7 +685,7 @@ void DTReadOutMapping::cacheMap() const {
       chanKey.clear();
       chanKey.push_back( link.dduId );
       chanKey.push_back( link.rosId );
-      theCache->rgDDU->insert( chanKey.begin(), chanKey.end(), entryNum );
+      localCache->rgDDU.insert( chanKey.begin(), chanKey.end(), entryNum );
 
       int mapId = link.cellId;
       std::vector<int>* dduMLgr;
@@ -725,12 +731,12 @@ void DTReadOutMapping::cacheMap() const {
         cellKey.push_back(  wheelId );
         cellKey.push_back( sectorId );
         std::vector<int>* dduMLgr = 0;
-        theCache->grDDU->find( cellKey.begin(), cellKey.end(), dduMLgr );
+        localCache->grDDU.find( cellKey.begin(), cellKey.end(), dduMLgr );
         if ( dduMLgr == 0 ) {
           std::unique_ptr<std::vector<int> > newVector(new std::vector<int>);
           dduMLgr = newVector.get();
-          theCache->grDDU->insert( cellKey.begin(), cellKey.end(),
-                                   std::move(newVector) );
+          localCache->grDDU.insert( cellKey.begin(), cellKey.end(),
+                                    std::move(newVector) );
         }
         dduMLgr->push_back( ientry );
       }
@@ -745,11 +751,11 @@ void DTReadOutMapping::cacheMap() const {
       delete dduEntry.second;
     }
 
-    theCache->rgBuf->clear();
-    theCache->grBuf->clear();
+    localCache->rgBuf.clear();
+    localCache->grBuf.clear();
   }
 
-  atomicCache().set(std::move(theCache));
+  atomicCache().set(std::move(localCache));
 
   return;
 }
